fix bonus files

+++ src_bonus/init_world/is_valid_line_list/skips/skip_value_bonus.c
@@ -1,17 +1,17 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   skip_value_bonus.c                                 :+:      :+:    :+:   */
+/*   skip_value.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: khanadat <khanadat@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/10 18:30:53 by ikawamuk          #+#    #+#             */
-/*   Updated: 2026/03/08 18:31:30 by khanadat         ###   ########.fr       */
+/*   Updated: 2026/03/15 20:17:36 by khanadat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "result_bonus.h"
-#include <float.h>
+#include "rt_define_bonus.h"
 #include <math.h>
 #include <stddef.h>
 
@@ -35,5 +35,5 @@ t_result	skip_sharp_angle(const char *line, size_t *line_idx)
 
 t_result	skip_length(const char *line, size_t *line_idx)
 {
-	return (skip_range(line, line_idx, DBL_EPSILON, INFINITY));
+	return (skip_range(line, line_idx, FLT_EPSILON, INFINITY));
 }
